src/app/app.component.spec.ts
import { async, ComponentFixture, TestBed } from '@angular/core/testing';

import { AppComponent } from './app.component';
import { HeaderComponent } from './header/header.component';
import { ActionComponent } from './action/action.component';
import { ListItemComponent } from './list-item/list-item.component';

describe('AppComponent', () => {

  let fixture: ComponentFixture<AppComponent>;

  beforeEach( async(() => {
    TestBed.configureTestingModule({
      declarations: [ AppComponent, HeaderComponent, ActionComponent, ListItemComponent ],
    }).compileComponents();
  }));

  beforeEach(() => {
    fixture = TestBed.createComponent(AppComponent);
  });

  it('should create the app', () => {
    const app = fixture.debugElement.componentInstance;
    expect(app).toBeTruthy();
  });

  it(`should have as title 'golf-ui'`, () => {
    const app = fixture.debugElement.componentInstance;
    expect(app.title).toEqual('golf-ui');
  });

  it('should render title in a h1 tag', () => {
    fixture.detectChanges();
    const compiled = fixture.debugElement.nativeElement;
    expect(compiled.querySelector('h1').textContent).toContain('golf-ui');
  });


});
